requirements.txt
python-dotenv
pymongo
pymongo[srv]
numpy
nltk
tensorflow